Remove ExponentialBackoff from rchttp client
There is only one type of retry method and that is exponential backoff.
Now the library will only retry with exponential backoff if maxRetries
is set to more than 0 (previously this would need to be set to at least 1,
otherwise the http response would result in a nil value with no error for
a 500 response from client.